requirements.txt
EasyProcess==1.1
entrypoint2==1.1
MouseInfo==0.1.3
mss==9.0.1
Pillow==10.0.1
PyAutoGUI==0.9.54
PyDirectInput==1.0.4
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
pytweening==1.0.7
